requirements.txt
requests
GitPython
schedule
